refactor feed scraping
This removes a lot of mutable state when it comes to string slices, which should make the code slightly easier to follow.
Performance is also improved as now feeds that don't meet the minimum listener requirement are filtered immediately.

Finally, scraping will now continue if an error is encountered while processing a feed. This removes virtually all information that could be available for
debugging should Broadcastify change their website, but any changes that would break the parser would likely require the entire thing to be rewritten anyway.