restructure main event loop to allow for destructors to be called when the program receives SIGINT / SIGTERM
This is mostly for SQLite so it can be closed cleanly without resorting to unsafe code.

Closing the SQLite connection during SIGINT / SIGTERM using unsafe is much easier (simply using mem::transmute and ptr::drop_in_place), but
introduces the ever so small chance of closing the connection in the middle of an update, among other fun potential corruption cases.