use master location list for feeds
This allows the user to specify string-based names for locations in the configuration file, rather than having to go fetch the ID for it on Broadcastify's website.
Also added is more specific location abbreviations to specify which country the feed is from. For example, before a feed from California would just show its location as "CA", but it will now
show as "US-CA".

Finally, this also renames the "state_id" field in the configuration file to be "process_location". The value should now be in a format like so: "<country>-<state/province/territory>".
For example, California would appear like this: "us-california", New York like this: "us-new-york", and Austrailia's New South Wales like this: "au-new-south-wales".

This does introduce the limitation that only feeds from this master list will ever be parsed, and any unknown ones will cause feed parsing to fail. Currently, all locations on Broadcastify
that have either a significant feed presence or airport feeds are on the list. Airport feeds are used as a factor for being on this list because there's a chance of a major event
occuring on them, at least more so than just amateur radio stations.